libre.sh adaptation + custom domain and mail config
This commit introduces an opinionated network setup due to the use
of libre.sh and an automatic reverse proxy.

A frontend.env file would contain environmental variables to
configure the external proxy.

VIRTUAL_HOST=pretix.ecobytes.net and VIRTUAL_PORT=80 need to be set.
LETSENCRYPT_HOST=-"- and LETSENCRYPT_EMAIL= can be set for Auto-TLS.

It further adds injection of custom configuration state and shows
an example for mail and domain configuration.